requirements.txt
Django==3.0.4
psycopg2==2.8.4
psycopg2-binary==2.8.4
django-extensions==2.2.8
djangorestframework==3.11.0
pytz==2019.3

pytest
pytest-mock
